Move homepage react code into its own file
